Moving the hometitle component
